use the same plugin for all the tests
